pyproject.toml
[build-system]
requires = ["setuptools>=69.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dd-log-analyzer"
version = "0.1.0"
description = "A CLI agent that analyzes Datadog logs — pattern detection, anomaly surfacing, error correlation, and trend analysis with Slack/Jira alerting."
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Jalil Abdollahi"},
]

dependencies = [
    "datadog-api-client>=2.20.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "numpy>=1.26.0",
    "httpx>=0.27.0",
    "aiosqlite>=0.20.0",
]

[project.optional-dependencies]
aws = [
    "boto3>=1.34.0",
]
web = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "PyJWT>=2.8.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "respx>=0.21",
    "ruff>=0.2.0",
    "moto[dynamodb,s3,ssm,secretsmanager]>=5.0",
    "boto3>=1.34.0",
]

[project.scripts]
dd-logs = "dd_log_analyzer.cli:cli"
dd-logs-web = "dd_log_analyzer.webapp.run:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
